score

+++ Assets/Scripts/GameManager.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using TMPro;
 using UnityEngine;
 using UnityEngine.SceneManagement;
+using UnityEngine.SocialPlatforms.Impl;
 
 public enum GameState
 {
@@ -25,7 +26,7 @@ public class GameManager : MonoBehaviour
     public int money;
     public int goal;
     public int day;
-    public float score;
+    public int score;
     public float groupedCarMoneyAmount;
     public TrafficLightsState currentTrafficLight;
     public List<Sprite> reactionSprites;
@@ -33,13 +34,13 @@ public class GameManager : MonoBehaviour
     public List<Sprite> trafficLightsUISprites;
     public CarMovement cars;
     public bool onMinigame = false;
-    
+
     public float greenLightDuration = 10f;
     public float yellowLightDuration = 3f;
     public float redLightDuration = 10f;
     public float countdown;
     private TrafficLightsState _lastTrafficLight;
-    
+
     private void Start()
     {
         if (Instance == null) Instance = this;
@@ -58,13 +59,13 @@ public class GameManager : MonoBehaviour
     private void LoadState(Scene scene, LoadSceneMode mode)
     {
         if (!PlayerPrefs.HasKey("SaveState")) return;
-        
+
         var data = PlayerPrefs.GetString("SaveState").Split('|');
         money = int.Parse(data[0]);
         day = int.Parse(data[1]);
-        score = float.Parse(data[2]);
+        score = int.Parse(data[2]);
         groupedCarMoneyAmount = float.Parse(data[3]);
-        currentTrafficLight = (TrafficLightsState) int.Parse(data[4]);
+        currentTrafficLight = (TrafficLightsState)int.Parse(data[4]);
         countdown = float.Parse(data[5]);
         reactionIndex = int.Parse(data[6]);
     }
